need further testing, but included an endpoint which needs to be call to call the AI model server
