+++ .env
@@ -1,2 +1,3 @@
 SECRET_KEY=bae80a0b7a6421d0c2d023c942d2087ff60947e9f891e3b644fdb4898b816793
 JWT_SECRET_KEY=7e0d63cfdf63822915d65c6dcd963203e1a807136fe8945bde92adeaeb17ea8cf104b95d5c2b7bd9f3cf393f6f26936d2eb8911e7b59ab16a9f6dd97d19abf45
+AI_SERVER_VALID_API_KEY=417f51ade292cb53d629081b357f8f1f58904007a77f85ab87d0d6b4a608e732
