+++ config.py
@@ -1,19 +1,16 @@
-import uuid
-import base64
-import os
 
+import os
 from flask import Flask, request, render_template
 from flask_cors import CORS
-from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
-from flask_login import LoginManager, login_user, logout_user, current_user
-from flask_restful import Api, Resource
+from flask_jwt_extended import JWTManager
+from flask_login import LoginManager
+from flask_restful import Api
 from flask_socketio import SocketIO, join_room, emit, send, disconnect
 
 from classes.users import users_bp, User
 from classes.videos import videos_bp
 from db import DatabaseManager
-
-
+from user_auth import DeviceUUID, Login, Status, Logout, QRLogin
 
 app = Flask(__name__)
 
@@ -27,7 +24,10 @@ def before_request(): # whenever the app is on, the database connection is open
 @app.teardown_request
 def teardown_request(exception):
     dbManager.stop()
-    app.logger.error(exception)
+    if exception:
+        app.logger.error(exception)
+    else:
+        app.logger.info('Database connection closed.')
 
 # Blueprint registrations
 app.register_blueprint(users_bp)
@@ -54,58 +54,6 @@ def load_user(user_id):
         return User(id=user['id'], username=user['username'], password=user['password'])
     return None
 
-class DeviceUUID(Resource):
-    def get(self):
-        random_uuid = uuid.uuid4()
-        return {'uuid': str(random_uuid)}
-
-class Login(Resource):
-    def post(self):
-
-        data = request.get_json()
-        username = data['username']
-        password = data['password']
-        projector_app_setting = data.get('projector_app_setting')
-
-        if not projector_app_setting:
-            video_link = 'retrieved from server action?'
-
-        user = dbManager.execute('SELECT * FROM users WHERE username = ?', (username,)).fetchone()
-
-        if user and (user['password'] == password):
-            user_obj = User(id=user['id'], username=user['username'], password=user['password'])
-            login_user(user_obj)
-
-            # Create a token for the authenticated user
-            access_token = create_access_token(identity=user['id'])
-            print(access_token)
-            return {'message': 'Logged in successfully', 'user_id': user['id'], 'token': access_token}, 200
-        return {'error': 'Invalid credentials'}, 401
-
-# Leo: changed to JWT
-class Status(Resource):
-    @jwt_required()
-    def get(self):
-        user_id = get_jwt_identity()
-
-        user = dbManager.execute('SELECT * FROM users WHERE id = ?', (user_id,)).fetchone()
-
-        if user:
-            return {
-                'logged_in': True,
-                'username': user['username'],
-                'password': user['password'],
-                'user_id': user['id']
-            }, 200
-        else:
-            return {'logged_in': False}, 401
-
-#Leo: I think the logout_user function is only for flask-login which is session, but not JWT
-class Logout(Resource):
-    @jwt_required()
-    def post(self):
-        logout_user()
-        return {'message': 'Logged out successfully'}, 200
 
 # Setup API
 api = Api(app)
@@ -113,46 +61,6 @@ api.add_resource(DeviceUUID, '/uuid')
 api.add_resource(Login, '/login')
 api.add_resource(Status, '/status')
 api.add_resource(Logout, '/logout')
-
-#code below added by Leo
-# u may tidy up
-class QRLogin(Resource):
-    @jwt_required()
-    def post(self):
-        data = request.get_json()
-        if not data or 'device_uuid' not in data:
-            return {'error': 'Invalid request'}, 400
-
-        user_id = get_jwt_identity()
-        device_uuid = data['device_uuid']
-        room = f'device_{device_uuid}'
-
-        if room not in socketio.server.manager.rooms.get('/', {}):
-            return {'error': 'Invalid device'}, 400
-        access_token = create_access_token(identity=user_id)
-
-        emit('QRLogin', {'login_success': 'true', 'user_id': user_id, 'token': access_token}, room=room, namespace='/')
-        return 200
-
-    @socketio.on('QRLogin')
-    def QRLogin_socketIO(data):
-        print(data)
-        if not data or 'device_uuid' not in data:
-            return
-        device_uuid = data['device_uuid']
-        room = f'device_{device_uuid}'
-        join_room(room)
-        emit('QRLogin', {'uuid': device_uuid, 'login_success': 'false'}, room=room)
-
-    @socketio.on('SyncSetting')
-    @jwt_required()
-    def SyncSetting_socketIO(data):
-        print(data)
-        user_id = get_jwt_identity()
-        room = f'room_{user_id}'
-        join_room(room)
-        emit('SyncSetting', data, room=room)
-
 api.add_resource(QRLogin, '/QRLogin')
 #---------until here------------
 
@@ -178,7 +86,6 @@ def handle_login(data):
 
 # Run the server
 if __name__ == '__main__':
-    init_db()
     # Path to your SSL certificate and key files
     ssl_context = ('./ssl_dynabot/cert.cert', './ssl_dynabot/key.key')
     socketio.run(app, host='0.0.0.0', port=8000, debug=True, ssl_context=ssl_context)
